Omnicia Audit Fix: ARD-01C  (#93)
* fix(ard-01c): update latestvotingpower var in grant to be type uint256

* fix(ard-01c): rebase fixes

+++ contracts/libraries/ARCDVestingVaultStorage.sol
@@ -33,7 +33,7 @@ library ARCDVestingVaultStorage {
         uint128 created;
         uint128 expiration;
         uint128 cliff;
-        uint128 latestVotingPower;
+        uint256 latestVotingPower;
         address delegatee;
     }
